backend/requirements.txt
# Core LLM providers
openai>=1.0.0
anthropic>=0.7.0
groq>=0.4.0
google-generativeai>=0.8.0
cerebras-cloud-sdk>=0.1.0

# Local LLM support
requests>=2.31.0

# FastAPI and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0

# Database and Supabase
psycopg2-binary>=2.9.0
supabase>=2.0.0

# ML and embeddings
sentence-transformers>=2.2.0
tqdm>=4.64.0
numpy>=1.24.0

# Existing dependencies
chromadb>=0.4.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0
python-dotenv>=1.0.0
streamlit>=1.28.0
pandas>=2.0.0
tiktoken>=0.5.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
fuzzywuzzy>=0.18.0
langdetect>=1.0.9

